+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- Attack vector for clickjacking.
 - Attack vector for CSRF.
 - Email address for users.
 - New HTML templates and forms.